backend/requirements.txt
# Core Django
Django>=4.2,<5.0
python-decouple>=3.8

# Django REST Framework
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
drf-yasg>=1.21.7

# Django Extensions
django-cors-headers>=4.3.0
django-filter>=23.5
django-mptt>=0.16.0
django-celery-beat>=2.5.0

# Database
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9

# Storage S3
django-storages>=1.14.0
boto3>=1.34.0

# Cache & Task Queue
redis>=5.0.1
celery>=5.3.4
django-redis>=5.4.0

# Static files
whitenoise>=6.6.0

# Media files
Pillow>=10.1.0

# Production server
gunicorn>=21.2.0

# Monitoring
sentry-sdk>=1.40.0

# Logging
python-json-logger>=2.0.7
boto3>=1.34.0
dj-database-url>=2.1.0
django-redis>=5.4.0
django-storages>=1.14.0
psycopg2-binary>=2.9.9
python-json-logger>=2.0.7
sentry-sdk>=1.40.0
boto3
celery
dj-database-url
Django
django-celery-beat
django-cors-headers
django-filter
django-mptt
django-redis
django-storages
djangorestframework
djangorestframework-simplejwt
drf-yasg
gunicorn
Pillow
psycopg2-binary
python-decouple
python-json-logger
redis
sentry-sdk
whitenoise
